Extract config builders into separate pkgs
