+++ internal/testutils/metric_pipeline_builder.go
@@ -1,4 +1,4 @@
-package metricpipeline
+package testutils
 
 import (
 	"fmt"
@@ -9,7 +9,7 @@ import (
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 )
 
-type metricPipelineBuilder struct {
+type MetricPipelineBuilder struct {
 	name              string
 	namespace         string
 	endpoint          string
@@ -19,8 +19,8 @@ type metricPipelineBuilder struct {
 	basicAuthPassword string
 }
 
-func newPipelineBuilder() *metricPipelineBuilder {
-	return &metricPipelineBuilder{
+func NewMetricPipelineBuilder() *MetricPipelineBuilder {
+	return &MetricPipelineBuilder{
 		name:           fmt.Sprintf("test-%d", time.Now().Nanosecond()),
 		namespace:      "telemetry-system",
 		endpoint:       "https://localhost",
@@ -29,28 +29,28 @@ func newPipelineBuilder() *metricPipelineBuilder {
 	}
 }
 
-func (b *metricPipelineBuilder) withName(name string) *metricPipelineBuilder {
+func (b *MetricPipelineBuilder) WithName(name string) *MetricPipelineBuilder {
 	b.name = name
 	return b
 }
 
-func (b *metricPipelineBuilder) withEndpoint(endpoint string) *metricPipelineBuilder {
+func (b *MetricPipelineBuilder) WithEndpoint(endpoint string) *MetricPipelineBuilder {
 	b.endpoint = endpoint
 	return b
 }
 
-func (b *metricPipelineBuilder) withRuntimeInputOn(on bool) *metricPipelineBuilder {
+func (b *MetricPipelineBuilder) WithRuntimeInputOn(on bool) *MetricPipelineBuilder {
 	b.runtimeInputOn = on
 	return b
 }
 
-func (b *metricPipelineBuilder) withBasicAuth(user, password string) *metricPipelineBuilder {
+func (b *MetricPipelineBuilder) WithBasicAuth(user, password string) *MetricPipelineBuilder {
 	b.basicAuthUser = user
 	b.basicAuthPassword = password
 	return b
 }
 
-func (b *metricPipelineBuilder) build() telemetryv1alpha1.MetricPipeline {
+func (b *MetricPipelineBuilder) Build() telemetryv1alpha1.MetricPipeline {
 	return telemetryv1alpha1.MetricPipeline{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      b.name,
